Refactor: move analytics/weather services to new namespaces
Significantly reorganized service interfaces and implementations for analytics and weather under Zucchinimvc.Application.Services.Analytical and Zucchinimvc.Application.Services.Weather. Updated usings and namespace declarations across affected files. Cleaned up redundant usings and ensured Program.cs files register services from new locations. This improves code structure, separation of concerns, and maintainability.
