+++ ZucchiniBackgroundJobs/Program.cs
@@ -1,12 +1,13 @@
+using Infrastrcture.Repositories;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Builder;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using Zucchinimvc.Application.Services.Weather;     
-using Zucchinimvc.Models.Entities;
-using Zucchinimvc.Repositories;
+using ZucchiniCore.Entities;
+using Zucchinimvc.Application.Services.Weather;
+
 
 var builder = FunctionsApplication.CreateBuilder(args);
