+++ ZucchiniMVC/Controllers/AnalyticsController.cs
@@ -1,14 +1,14 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using Zucchinimvc.Models.ViewModels;
-using Zucchinimvc.Services;
-using Zucchinimvc.Services.API;
+using Zucchinimvc.Application.Services.Analytical;
+
+
 
 namespace Zucchinimvc.Controllers
 {
     public class AnalyticsController : Controller
     {
         private readonly IAnalyticsService _analyticsService;
-    
+
         public AnalyticsController(IAnalyticsService analyticsService)
         {
             _analyticsService = analyticsService;
